refactored repo
