+++ assembly/Fraction.ts
@@ -1,24 +1,23 @@
 import { BigInt } from "as-bigint";
 import { BigNumber, Rounding } from "as-bignumber";
 
-export { Rounding } from "as-bignumber";
+// TODO: check for overflow and throw
 
-export class Fraction {
-  public readonly numerator: BigInt;
-  public readonly denominator: BigInt;
+export class Fraction<T extends Number> {
+  public readonly numerator: T;
+  public readonly denominator: T;
 
   // 155, the number of digits in the maximum value of a 512 bit integer
   public static DEFAULT_PRECISION: i32 = 155;
   public static DEFAULT_ROUNDING: Rounding = Rounding.HALF_UP;
-  public static readonly MAX_POWER: i32 = 999999999;
-
-  public static readonly ONE: Fraction = new Fraction(BigInt.ONE);
-  public static readonly HALF: Fraction = new Fraction(BigInt.ONE, BigInt.fromUInt16(2));
 
   // CONSTRUCTORS //////////////////////////////////////////////////////////////////////////////////////////////////////
 
-  constructor(numerator: BigInt, denominator: BigInt = BigInt.ONE) {
-    if (denominator.isZero()) {
+  constructor(numerator: T, denominator: T) {
+    if (!isInteger<T>(denominator)) {
+      throw new Error("Fraction numerator and denominator must be integers");
+    }
+    if (denominator == 0) {
       throw new Error("Divide by zero");
     }
     this.numerator = numerator;
@@ -31,79 +30,114 @@ export class Fraction {
    * @param  val the number as {BigNumber}, {BigInt}, {BigInt[]}, string, or {number}
    * @return the new {Fraction} instance
    */
-  static from<T>(val: T): Fraction {
-    if (val instanceof Fraction) return val;
+  static from<TOut extends Number, TArg>(val: TArg): Fraction<TOut> {
+    if (val instanceof Fraction) {
+      if (nameof<TOut>() == nameof(val.numerator)) return val;
+      return val.toFraction<TOut>();
+    }
     // @ts-ignore
-    if (val instanceof Array) return Fraction.fromArray(val);
+    if (isArray<TArg>()) return this.fromArray<TOut>(val);
     // @ts-ignore
-    if (val instanceof BigNumber) return Fraction.fromBigNumber(val);
+    if (val instanceof BigNumber) return Fraction.fromBigNumber<TOut>(val);
     // @ts-ignore
-    if (val instanceof string) return Fraction.fromString(val);
+    if (isString<TArg>()) return Fraction.fromString<TOut>(val);
     // @ts-ignore
-    if (val instanceof BigInt) return new Fraction(val);
+    if (val instanceof BigInt) return Fraction.fromBigInt<TOut>(val);
     // @ts-ignore
-    if (val instanceof f32) return Fraction.fromBigNumber(BigNumber.fromFloat64(<f64>val));
+    if (isFloat<TArg>()) return Fraction.fromBigNumber<TOut>(BigNumber.fromFloat64(<f64>val));
     // @ts-ignore
-    if (val instanceof f64) return Fraction.fromBigNumber(BigNumber.fromFloat64(val));
-    // @ts-ignore
-    if (val instanceof i8) return new Fraction(BigInt.fromInt16(<i16>val));
-    // @ts-ignore
-    if (val instanceof u8) return new Fraction(BigInt.fromUInt16(<u16>val));
-    // @ts-ignore
-    if (val instanceof i16) return new Fraction(BigInt.fromInt16(val));
-    // @ts-ignore
-    if (val instanceof u16) return new Fraction(BigInt.fromUInt16(val));
-    // @ts-ignore
-    if (val instanceof i32) return new Fraction(BigInt.fromInt32(val));
-    // @ts-ignore
-    if (val instanceof u32) return new Fraction(BigInt.fromUInt32(val));
-    // @ts-ignore
-    if (val instanceof i64) return new Fraction(BigInt.fromInt64(val));
-    // @ts-ignore
-    if (val instanceof u64) return new Fraction(BigInt.fromUInt64(val));
+    if (isInteger<TArg>()) return new Fraction(val, 1).toFraction<TOut>();
 
-    throw new TypeError("Unsupported generic type " + nameof<T>(val));
+    throw new TypeError("Unsupported generic type " + nameof<TArg>(val));
   }
 
-  static fromArray<TInt>(arr: TInt[]): Fraction {
+  static fromArray<TOut extends Number, TArg extends Number>(arr: Array<TArg>): Fraction<TOut> {
+    if (!isInteger<TOut>()) {
+      throw new TypeError("Unsupported generic type " + nameof<TOut>());
+    }
+    if (!isInteger<TArg>(arr[0])) {
+      throw new TypeError("Unsupported generic type " + nameof<TArg>(arr[0]));
+    }
+    if (arr.length != 2) {
+      throw new Error(
+        "Unexpected length: expected an array of length 2 in the form [numerator, denominator]"
+      );
+    }
+    return new Fraction(<TOut><unknown>arr[0], <TOut><unknown>arr[1]);
+  }
+
+  static fromArrayToInferredType<TArg extends Number, TOut extends Number = TArg>(arr: TArg[]): Fraction<TOut> {
+    if (!isInteger<TOut>()) {
+      throw new TypeError("Unsupported generic type " + nameof<TOut>());
+    }
+    if (!isInteger<TArg>(arr[0])) {
+      throw new TypeError("Unsupported generic type " + nameof<TArg>(arr[0]));
+    }
     if (arr.length != 2) {
       throw new Error(
           "Unexpected length: expected an array of length 2 in the form [numerator, denominator]"
       );
     }
-    return new Fraction(BigInt.from(arr[0]), BigInt.from(arr[1]));
+    return new Fraction(<TOut><unknown>arr[0], <TOut><unknown>arr[1]);
   }
-
-  static fromBigNumber(val: BigNumber): Fraction {
+  
+  static fromBigNumber<TOut extends Number = i64>(val: BigNumber, unsafe = false): Fraction<TOut> {
+    if (!isInteger<TOut>()) {
+      throw new TypeError("Unsupported generic type " + nameof<TOut>());
+    }
+    if (!unsafe) {
+      const outType: string = nameof<TOut>();
+      const precision: i32 = val.precision;
+      Fraction.checkBigNumberSafety(outType, precision);
+    }
     if (val.e > 0) {
       const scale = BigInt.fromString("1".padEnd(1 + val.e, "0"));
-      return new Fraction(val.m.copy(), scale);
+      if (val.m.isNegative) {
+        return new Fraction(<TOut><unknown>val.m.toInt64(), <TOut><unknown>scale.toUInt32());
+      }
+      return new Fraction(<TOut><unknown>val.m.toUInt64(), <TOut><unknown>scale.toUInt32());
     } else if (val.e < 0) {
       const scale = BigInt.fromString("1".padEnd(1 - val.e, "0"));
-      return new Fraction(val.m.mul(scale));
+      const intVal: BigInt = val.m.mul(scale);
+      if (intVal.isNegative) {
+        return new Fraction(<TOut><unknown>intVal.toInt64(), <TOut><unknown>1);
+      }
+      return new Fraction(<TOut><unknown>intVal.toUInt64(), <TOut><unknown>1);
     } else {
-      return new Fraction(val.m.copy());
+      if (val.m.isNegative) {
+        return new Fraction(<TOut><unknown>val.m.toInt64(), <TOut><unknown>1);
+      }
+      return new Fraction(<TOut><unknown>val.m.toUInt64(), <TOut><unknown>1);
     }
   }
 
-  static fromBigInt(val: BigInt): Fraction {
-    return new Fraction(val.copy());
+  static fromBigInt<TOut extends Number = i64>(val: BigInt): Fraction<TOut> {
+    if (!isInteger<TOut>()) {
+      throw new TypeError("Unsupported generic type " + nameof<TOut>());
+    }
+    if (val.isNegative) {
+      return new Fraction(<TOut><unknown>val.toInt64(), <TOut><unknown>1);
+    }
+    return new Fraction(<TOut><unknown>val.toUInt64(), <TOut><unknown>1);
   }
   
-  static fromString(val: string): Fraction {
-    return Fraction.fromBigNumber(BigNumber.fromString(val, 0));
+  static fromString<TOut extends Number = i64>(val: string): Fraction<TOut> {
+    if (!isInteger<TOut>()) {
+      throw new TypeError("Unsupported generic type " + nameof<TOut>());
+    }
+    return Fraction.fromBigNumber<TOut>(BigNumber.fromString(val, 0));
   }
 
   // OUTPUT ////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
   toString(): string {
-    return `[${this.numerator.toString()}, ${this.denominator.toString()}]`
+    return `[${this.numerator}, ${this.denominator}]`
   }
 
   toNumberString(precision: i32 = Fraction.DEFAULT_PRECISION, rounding: Rounding = Fraction.DEFAULT_ROUNDING): string {
     return BigNumber.fromFraction(
-        this.numerator,
-        this.denominator,
+        BigInt.from(this.numerator),
+        BigInt.from(this.denominator),
         precision,
         rounding
     ).toString();
@@ -117,8 +151,8 @@ export class Fraction {
       return "0";
     }
     return BigNumber.fromFraction(
-        this.numerator,
-        this.denominator,
+        BigInt.from(this.numerator),
+        BigInt.from(this.denominator),
         Fraction.DEFAULT_PRECISION,
         rounding
     ).toSignificant(digits, rounding);
@@ -129,8 +163,8 @@ export class Fraction {
       rounding: Rounding = Rounding.HALF_UP
   ): string {
     return BigNumber.fromFraction(
-        this.numerator,
-        this.denominator,
+        BigInt.from(this.numerator),
+        BigInt.from(this.denominator),
         Fraction.DEFAULT_PRECISION,
         rounding
     ).toFixed(places, rounding);
@@ -144,287 +178,399 @@ export class Fraction {
 
   // performs floor division
   toBigInt(): BigInt {
-    return this.numerator.div(this.denominator);
+    // @ts-ignore
+    return BigInt.from(this.numerator / this.denominator);
   }
 
-  toArray(): BigInt[] {
-    return [this.numerator.copy(), this.denominator.copy()];
+  toArray<TOut extends Number = T>(): TOut[] {
+    return [<TOut><unknown>this.numerator, <TOut><unknown>this.denominator];
+  }
+
+  toFraction<TOut extends Number>(): Fraction<TOut> {
+    return new Fraction(<TOut><unknown>this.numerator, <TOut><unknown>this.denominator);
   }
 
   quotient(precision: i32 = Fraction.DEFAULT_PRECISION, rounding: Rounding = Fraction.DEFAULT_ROUNDING): BigNumber {
     return this.toBigNumber(precision, rounding);
   }
 
-  // O(N)
-  copy(): Fraction {
-    return new Fraction(this.numerator.copy(), this.denominator.copy());
+  copy<TOut extends Number = T>(): Fraction<TOut> {
+    return new Fraction(<TOut><unknown>this.numerator, <TOut><unknown>this.denominator);
   }
 
-  // O(N)
-  opposite(): Fraction  {
-    return new Fraction(this.numerator.opposite(), this.denominator.copy());
+  opposite<TOut extends Number = T>(): Fraction<TOut> {
+    // @ts-ignore
+    return new Fraction<TOut>(-1 * <TOut><unknown>this.numerator, <TOut><unknown>this.denominator);
+  }
+  
+  abs<TOut extends Number = T>(): Fraction<TOut> {
+    // @ts-ignore
+    return new Fraction(<TOut><unknown>abs(this.numerator), <TOut><unknown>abs(this.denominator));
   }
 
-  // O(N)
-  abs(): Fraction  {
-    return new Fraction(this.numerator.abs(), this.denominator.abs());
-  }
-
-  reciprocal(): Fraction {
-    return new Fraction(this.denominator.copy(), this.numerator.copy());
+  reciprocal<TOut extends Number = T>(): Fraction<TOut> {
+    return new Fraction(<TOut><unknown>this.denominator, <TOut><unknown>this.numerator);
   }
 
   // COMPARISON OPERATORS //////////////////////////////////////////////////////////////////////////////////////////////
 
-  eq<T>(other: T): boolean {
+  eq<U>(other: U): boolean {
     return this.compareTo(Fraction.from(other)) == 0;
   }
 
-  ne<T>(other: T): boolean {
+  ne<U>(other: U): boolean {
     return !this.eq(Fraction.from(other));
   }
 
-  lt<T>(other: T): boolean {
+  lt<U>(other: U): boolean {
     return this.compareTo(Fraction.from(other)) < 0;
   }
 
-  lte<T>(other: T): boolean {
+  lte<U>(other: U): boolean {
     return this.compareTo(Fraction.from(other)) <= 0;
   }
 
-  gt<T>(other: T): boolean {
+  gt<U>(other: U): boolean {
     return this.compareTo(Fraction.from(other)) > 0;
   }
 
-  gte<T>(other: T): boolean {
+  gte<U>(other: U): boolean {
     return this.compareTo(Fraction.from(other)) >= 0;
   }
 
-  compareTo(other: Fraction): i32 {
-    const leftIsNeg: boolean = this.isNegative;
-    const rightIsNeg: boolean = other.isNegative;
-    if (leftIsNeg && !rightIsNeg) {
-      return -1;
-    } else if (!leftIsNeg && rightIsNeg) {
-      return 1;
+  compareTo<U>(other: U): i32 {
+    const right = Fraction.from(other);
+    if (isSigned(this.numerator) || isSigned(right.numerator)) {
+      const leftIsNeg: boolean = this.isNegative;
+      const rightIsNeg: boolean = right.isNegative;
+      if (leftIsNeg && !rightIsNeg) {
+        return -1;
+      } else if (!leftIsNeg && rightIsNeg) {
+        return 1;
+      }
     }
-    const a = BigInt.mul(this.numerator, other.denominator);
-    const b = BigInt.mul(other.numerator, this.denominator);
-    return leftIsNeg ? b.magCompareTo(a) : a.magCompareTo(b);
+    return this.magCompareTo(right);
   }
 
-  magCompareTo(other: Fraction): i32 {
-    const left = BigInt.mul(this.numerator, other.denominator);
-    const right = BigInt.mul(other.numerator, this.denominator);
-    return left.magCompareTo(right);
+  magCompareTo<U>(other: U): i32 {
+    const right = Fraction.from(other);
+    if (isSigned(this.numerator) || isSigned(right.numerator)) {
+      const a: i64 = <i64><unknown>this.numerator * <i64><unknown>right.denominator;
+      const b: i64 = <i64><unknown>right.numerator * <i64><unknown>this.denominator;
+      const absA: i64 = abs(a);
+      const absB: i64 = abs(b);
+      return this.isNegative ? absB - absA : absA - absB;
+    }
+    const a: u64 = <u64><unknown>this.numerator * <u64><unknown>right.denominator;
+    const b: u64 = <u64><unknown>right.numerator * <u64><unknown>this.denominator;
+    if (this.isNegative) {
+      if (b < a) return -1;
+      if (b > a) return 1;
+      return 0;
+    } else {
+      if (a < b) return -1;
+      if (a > b) return 1;
+      return 0;
+    }
   }
 
   // ARITHMETIC ////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-  add<T>(other: T): Fraction {
+  add<U, TOut extends Number = T>(other: U): Fraction<TOut> {
+    if (!isInteger<TOut>()) {
+      throw new TypeError("Unsupported generic type " + nameof<TOut>());
+    }
     const addend = Fraction.from(other);
-    if (this.denominator.eq(addend.denominator)) {
-      const numeratorSum: BigInt = BigInt.add(this.numerator, addend.numerator);
-      return new Fraction(numeratorSum, this.denominator);
+    if (isSigned(this.numerator) || isSigned(addend.numerator)) {
+      const thisNum: i64 = <i64><unknown>this.numerator;
+      const thisDenom: i64 = <i64><unknown>this.denominator;
+      const otherNum: i64 = <i64><unknown>addend.numerator;
+      const otherDenom: i64 = <i64><unknown>addend.denominator;
+      if (thisDenom == otherDenom) {
+        return new Fraction(<TOut><unknown>(thisNum + otherNum), <TOut><unknown>thisDenom);
+      }
+      return new Fraction(
+        <TOut><unknown>(thisNum * otherDenom + otherNum * thisDenom),
+        <TOut><unknown>(thisDenom * otherDenom)
+      );
+    }
+    const thisNum: u64 = <u64><unknown>this.numerator;
+    const thisDenom: u64 = <u64><unknown>this.denominator;
+    const otherNum: u64 = <u64><unknown>addend.numerator;
+    const otherDenom: u64 = <u64><unknown>addend.denominator;
+    if (thisDenom == otherDenom) {
+      return new Fraction(<TOut><unknown>(thisNum + otherNum), <TOut><unknown>thisDenom);
     }
     return new Fraction(
-        BigInt.add(
-            BigInt.mul(this.numerator, addend.denominator),
-            BigInt.mul(addend.numerator, this.denominator)
-        ),
-        BigInt.mul(this.denominator, addend.denominator)
+      <TOut><unknown>(thisNum * otherDenom + otherNum * thisDenom),
+      <TOut><unknown>(thisDenom * otherDenom)
     );
   }
 
-  sub<T>(other: T): Fraction {
-    const subtrahend: Fraction = Fraction.from(other);
-    if (BigInt.eq(this.denominator, subtrahend.denominator)) {
-      const numeratorSub: BigInt = BigInt.sub(this.numerator, subtrahend.numerator);
-      return new Fraction(numeratorSub, this.denominator);
+  sub<U, TOut extends Number = T>(other: U): Fraction<TOut> {
+    if (!isInteger<TOut>()) {
+      throw new TypeError("Unsupported generic type " + nameof<TOut>());
+    }
+    const subtrahend = Fraction.from(other);
+    if (isSigned(this.numerator || isSigned(subtrahend.numerator))) {
+      const thisNum: i64 = <i64><unknown>this.numerator;
+      const thisDenom: i64 = <i64><unknown>this.denominator;
+      const otherNum: i64 = <i64><unknown>subtrahend.numerator;
+      const otherDenom: i64 = <i64><unknown>subtrahend.denominator;
+      if (thisDenom == otherDenom) {
+        return new Fraction(<TOut><unknown>(thisNum - otherNum), <TOut><unknown>thisDenom);
+      }
+      return new Fraction(
+        <TOut><unknown>(thisNum * otherDenom - otherNum * thisDenom),
+        <TOut><unknown>(thisDenom * otherDenom)
+      );
+    }
+    const thisNum: u64 = <u64><unknown>this.numerator;
+    const thisDenom: u64 = <u64><unknown>this.denominator;
+    const otherNum: u64 = <u64><unknown>subtrahend.numerator;
+    const otherDenom: u64 = <u64><unknown>subtrahend.denominator;
+    if (thisDenom == otherDenom) {
+      return new Fraction(<TOut><unknown>(thisNum - otherNum), <TOut><unknown>thisDenom);
     }
     return new Fraction(
-        BigInt.sub(
-            BigInt.mul(this.numerator, subtrahend.denominator),
-            BigInt.mul(subtrahend.numerator, this.denominator)
-        ),
-        BigInt.mul(this.denominator, subtrahend.denominator)
+      <TOut><unknown>(thisNum * otherDenom - otherNum * thisDenom),
+      <TOut><unknown>(thisDenom * otherDenom)
     );
   }
 
-  mul<T>(other: T): Fraction {
-    const multiplier: Fraction = Fraction.from(other);
-    return new Fraction(
-        BigInt.mul(this.numerator, multiplier.numerator),
-        BigInt.mul(this.denominator, multiplier.denominator)
+  mul<U, TOut extends Number = T>(other: U): Fraction<TOut> {
+    if (!isInteger<TOut>()) {
+      throw new TypeError("Unsupported generic type " + nameof<TOut>());
+    }
+    const multiplier = Fraction.from(other);
+    return new Fraction<TOut>(
+      // @ts-ignore
+      <TOut><unknown>this.numerator * <TOut><unknown>multiplier.numerator,
+      // @ts-ignore
+      <TOut><unknown>this.denominator * <TOut><unknown>multiplier.denominator
     );
   }
 
-  square(): Fraction {
-    return new Fraction(
-        this.numerator.square(),
-        this.denominator.square()
+  square<TOut extends Number = T>(): Fraction<TOut> {
+    if (!isInteger<TOut>()) {
+      throw new TypeError("Unsupported generic type " + nameof<TOut>());
+    }
+    return new Fraction<TOut>(
+      // @ts-ignore
+      <TOut><unknown>this.numerator * <TOut><unknown>this.numerator,
+      // @ts-ignore
+      <TOut><unknown>this.denominator * <TOut><unknown>this.denominator
     );
   }
 
   /**
    * Divides two Fractions and rounds result
    */
-  div<T>(other: T): Fraction {
-    const divisor: Fraction = Fraction.from(other);
-    if (divisor.numerator.isZero()) {
+  div<U, TOut extends Number = T>(other: U): Fraction<TOut> {
+    if (!isInteger<TOut>()) {
+      throw new TypeError("Unsupported generic type " + nameof<TOut>());
+    }
+    const divisor = Fraction.from(other);
+    if (divisor.numerator == 0) {
       throw new Error("Divide by zero");
     }
-    return new Fraction(
-        BigInt.mul(this.numerator, divisor.denominator),
-        BigInt.mul(this.denominator, divisor.numerator)
+    return new Fraction<TOut>(
+      // @ts-ignore
+      <TOut><unknown>this.numerator * <TOut><unknown>divisor.denominator,
+      // @ts-ignore
+      <TOut><unknown>this.denominator * <TOut><unknown>divisor.numerator
     );
   }
 
-  sqrt(): Fraction {
-    if (this.isNegative)
-      throw new RangeError("Square root of negative numbers is not supported");
-    if (this.isZero()) {
-      return this.copy();
+  // TODO: why can't i take sqrt(numerator) / sqrt(denominator) ??
+  sqrt<TOut extends Number = T>(): Fraction<TOut> {
+    if (!isInteger<TOut>()) {
+      throw new TypeError("Unsupported generic type " + nameof<TOut>());
     }
-    return Fraction.fromBigNumber(this.toBigNumber().sqrt());
+    if (this.isNegative) {
+      throw new RangeError("Square root of negative numbers is not supported");
+    }
+    // @ts-ignore
+    if (this.numerator == 0) {
+      return this.toFraction<TOut>();
+    }
+    const outType = nameof<TOut>();
+    if (outType == I64.name || outType == U64.name) {
+      return Fraction.fromBigNumber<TOut>(this.toBigNumber().sqrt(19), true);
+    } else if (outType == I32.name || outType == U32.name) {
+      return Fraction.fromBigNumber<TOut>(this.toBigNumber().sqrt(10), true);
+    } else if (outType == I16.name || outType == U16.name) {
+      return Fraction.fromBigNumber<TOut>(this.toBigNumber().sqrt(5), true);
+    } else {
+      return Fraction.fromBigNumber<TOut>(this.toBigNumber().sqrt(3), true);
+    }
   }
 
-  pow(k: i32): Fraction {
-    if (k < -Fraction.MAX_POWER || k > Fraction.MAX_POWER) {
-      throw new Error(`Power argument out of bounds [-${Fraction.MAX_POWER}, ${Fraction.MAX_POWER}]: ${k}`);
+  pow<TOut extends Number = T>(k: TOut): Fraction<TOut> {
+    if (!isInteger<TOut>(k)) {
+      throw new TypeError("Unsupported generic type " + nameof<TOut>(k));
     }
     if (k < 0) {
-      return new Fraction(this.denominator.pow(k), this.numerator.pow(k));
+      // @ts-ignore
+      return new Fraction<TOut>((<TOut>this.denominator) ** k, (<TOut>this.numerator) ** k);
     }
-    return new Fraction(this.numerator.pow(k), this.denominator.pow(k));
+    // @ts-ignore
+    return new Fraction<TOut>((<TOut>this.numerator) ** k, (<TOut>this.denominator) ** k);
   }
 
   // UTILITIES /////////////////////////////////////////////////////////////////////////////////////////////////////////
 
   get isNegative(): boolean {
-    return this.numerator.isNegative != this.denominator.isNegative;
+    // @ts-ignore
+    return this.numerator < 0 != this.denominator < 0;
   }
 
   isZero(): boolean {
-    return this.numerator.isZero();
+    // @ts-ignore
+    return this.numerator == 0;
   }
 
   get isInteger(): boolean {
-    return this.toBigNumber().isInteger;
+    // @ts-ignore
+    return this.numerator >= this.denominator && (this.numerator % this.denominator == 0)
+  }
+
+  private static checkBigNumberSafety(outType: string, precision: i32): void {
+    if (outType == I64.name || outType == U64.name) {
+      if (precision > 19) {
+        throw new Error(`Integer overflow: BigNumber input has precision of ${precision} digits while 64 bit integers fit only 19 digits`);
+      }
+    } else if (outType == I32.name || outType == U32.name) {
+      if (precision > 10) {
+        throw new Error (`Integer overflow: BigNumber input has precision of ${precision} digits while 32 bit integers fit only 10 digits`);
+      }
+    } else if (outType == I16.name || outType == U16.name) {
+      if (precision > 5) {
+        throw new Error (`Integer overflow: BigNumber input has precision of ${precision} digits while 16 bit integers fit only 5 digits`);
+      }
+    } else if (outType == I8.name || outType == U8.name) {
+      if (precision > 3) {
+       throw new Error (`Integer overflow: BigNumber input has precision of ${precision} digits while 8 bit integers fit only 3 digits`);
+     }
+    }
   }
 
   // SYNTAX SUGAR ///////////////////////////////////////////////////////////////////////////////////////////////////
 
   static eq<T, U>(left: T, right: U): boolean {
-    const a: Fraction = Fraction.from(left);
+    const a = Fraction.from(left);
     return a.eq(right);
   }
 
   @operator("==")
-  private static eqOp(left: Fraction, right: Fraction): boolean {
+  private static eqOp<TInt extends Number>(left: Fraction<TInt>, right: Fraction<TInt>): boolean {
     return left.eq(right);
   }
 
   static ne<T, U>(left: T, right: U): boolean {
-    const a: Fraction = Fraction.from(left);
+    const a = Fraction.from(left);
     return a.ne(right);
   }
 
   @operator("!=")
-  private static neOp(left: Fraction, right: Fraction): boolean {
+  private static neOp<TInt extends Number>(left: Fraction<TInt>, right: Fraction<TInt>): boolean {
     return left.ne(right);
   }
 
   static lt<T, U>(left: T, right: U): boolean {
-    const a: Fraction = Fraction.from(left);
+    const a = Fraction.from(left);
     return a.lt(right);
   }
 
   @operator("<")
-  private static ltOp(left: Fraction, right: Fraction): boolean {
+  private static ltOp<TInt extends Number>(left: Fraction<TInt>, right: Fraction<TInt>): boolean {
     return left.lt(right);
   }
 
   static lte<T, U>(left: T, right: U): boolean {
-    const a: Fraction = Fraction.from(left);
+    const a = Fraction.from(left);
     return a.lte(right);
   }
 
   @operator("<=")
-  private static lteOp(left: Fraction, right: Fraction): boolean {
+  private static lteOp<TInt extends Number>(left: Fraction<TInt>, right: Fraction<TInt>): boolean {
     return left.lte(right);
   }
 
   static gt<T, U>(left: T, right: U): boolean {
-    const a: Fraction = Fraction.from(left);
+    const a = Fraction.from(left);
     return a.gt(right);
   }
 
   @operator(">")
-  private static gtOp(left: Fraction, right: Fraction): boolean {
+  private static gtOp<TInt extends Number>(left: Fraction<TInt>, right: Fraction<TInt>): boolean {
     return left.gt(right);
   }
 
   static gte<T, U>(left: T, right: U): boolean {
-    const a: Fraction = Fraction.from(left);
+    const a = Fraction.from(left);
     return a.gte(right);
   }
 
   @operator(">=")
-  private static gteOp(left: Fraction, right: Fraction): boolean {
+  private static gteOp<TInt extends Number>(left: Fraction<TInt>, right: Fraction<TInt>): boolean {
     return left.gte(right);
   }
 
-  static add<T, U>(left: T, right: U): Fraction {
-    const a: Fraction = Fraction.from(left);
-    return a.add(right);
+  static add<TOut extends Number, T, U>(left: T, right: U): Fraction<TOut> {
+    const a = Fraction.from(left);
+    return a.add<U, TOut>(right);
   }
 
   @operator("+")
-  private static addOp(left: Fraction, right: Fraction): Fraction {
+  private static addOp<TInt extends Number>(left: Fraction<TInt>, right: Fraction<TInt>): Fraction<TInt> {
     return left.add(right);
   }
 
-  static sub<T, U>(left: T, right: U): Fraction {
-    const a: Fraction = Fraction.from(left);
-    return a.sub(right);
+  static sub<TOut extends Number, T, U>(left: T, right: U): Fraction<TOut> {
+    const a = Fraction.from(left);
+    return a.sub<U, TOut>(right);
   }
 
   @operator("-")
-  private static subOp(left: Fraction, right: Fraction): Fraction {
+  private static subOp<TInt extends Number>(left: Fraction<TInt>, right: Fraction<TInt>): Fraction<TInt> {
     return left.sub(right);
   }
 
-  static mul<T, U>(left: T, right: U): Fraction {
-    const a: Fraction = Fraction.from(left);
-    return a.mul(right);
+  static mul<TOut extends Number, T, U>(left: T, right: U): Fraction<TOut> {
+    const a = Fraction.from(left);
+    return a.mul<U, TOut>(right);
   }
 
   @operator("*")
-  private static mulOp(left: Fraction, right: Fraction): Fraction {
+  private static mulOp<TInt extends Number>(left: Fraction<TInt>, right: Fraction<TInt>): Fraction<TInt>{
     return left.mul(right);
   }
 
-  static div<T, U>(left: T, right: U): Fraction {
-    const a: Fraction = Fraction.from(left);
-    return a.div(right);
+  static div<TOut extends Number, T, U>(left: T, right: U): Fraction<TOut> {
+    const a = Fraction.from(left);
+    return a.div<U, TOut>(right);
   }
 
   @operator("/")
-  static divOp(left: Fraction, right: Fraction): Fraction {
+  static divOp<TInt extends Number>(left: Fraction<TInt>, right: Fraction<TInt>): Fraction<TInt> {
     return left.div(right);
   }
 
-  static pow<T>(base: T, k: i32): Fraction {
-    const x: Fraction = Fraction.from(base);
+  static pow<TOut extends Number, U>(base: U, k: TOut): Fraction<TOut> {
+    if (!isInteger<TOut>(k)) {
+      throw new TypeError("Unsupported generic type " + nameof<TOut>(k));
+    }
+    const x = Fraction.from(base);
     return x.pow(k);
   }
 
   @operator("**")
-  private static powOp(left: Fraction, right: Fraction): Fraction {
+  private static powOp<TInt extends  Number>(left: Fraction<TInt>, right: Fraction<TInt>): Fraction<TInt> {
     if (!right.isInteger) {
       throw new Error("Exponent must be an integer value");
     }
-    return left.pow(right.toBigInt().toInt32());
+    // @ts-ignore
+    return left.pow(<TInt><unknown>(right.numerator / right.denominator));
   }
 }
