+++ assembly/__tests__/BigFraction/arithmetic.spec.ts
@@ -1,4 +1,4 @@
-import { Fraction } from "../Fraction";
+import { BigFraction } from "../../BigFraction";
 import { TestCase } from "./testUtils/TestCase";
 import { plannedTestCases } from "./testUtils/testCases";
 import { randomTestCases } from "./testUtils/randomTestCases";
@@ -10,10 +10,10 @@ describe("Arithmetic operations", () => {
   it("adds", () => {
     for (let i = 0; i < testCases.length; i++) {
       const testCase: TestCase = testCases[i];
-      const x = Fraction.fromString(testCase.x);
-      const y = Fraction.fromString(testCase.y);
+      const x = BigFraction.fromString(testCase.x);
+      const y = BigFraction.fromString(testCase.y);
       const result = x.add(y).toFixed(30);
-      const expected = Fraction.fromString(testCase.sum).toFixed(30);
+      const expected = BigFraction.fromString(testCase.sum).toFixed(30);
       expect(result).toStrictEqual(expected);
     }
   });
@@ -21,10 +21,10 @@ describe("Arithmetic operations", () => {
   it("subtracts", () => {
     for (let i = 0; i < testCases.length; i++) {
       const testCase: TestCase = testCases[i];
-      const x = Fraction.fromString(testCase.x);
-      const y = Fraction.fromString(testCase.y);
+      const x = BigFraction.fromString(testCase.x);
+      const y = BigFraction.fromString(testCase.y);
       const result = x.sub(y).toFixed(30);
-      const expected = Fraction.fromString(testCase.difference).toFixed(30);
+      const expected = BigFraction.fromString(testCase.difference).toFixed(30);
       expect(result).toStrictEqual(expected);
     }
   });
@@ -32,10 +32,10 @@ describe("Arithmetic operations", () => {
   it("multiplies", () => {
     for (let i = 0; i < testCases.length; i++) {
       const testCase: TestCase = testCases[i];
-      const x = Fraction.fromString(testCase.x);
-      const y = Fraction.fromString(testCase.y);
+      const x = BigFraction.fromString(testCase.x);
+      const y = BigFraction.fromString(testCase.y);
       const result = x.mul(y).toFixed(30);
-      const expected = Fraction.fromString(testCase.product).toFixed(30);
+      const expected = BigFraction.fromString(testCase.product).toFixed(30);
       expect(result).toStrictEqual(expected);
     }
   });
@@ -43,13 +43,13 @@ describe("Arithmetic operations", () => {
   it("divides", () => {
     for (let i = 0; i < testCases.length; i++) {
       const testCase: TestCase = testCases[i];
-      const x = Fraction.fromString(testCase.x);
-      const y = Fraction.fromString(testCase.y);
+      const x = BigFraction.fromString(testCase.x);
+      const y = BigFraction.fromString(testCase.y);
       if (y.isZero()) {
         continue;
       }
       const result = x.div(y).toFixed(30);
-      const expected = Fraction.fromString(testCase.quotient).toFixed(30);
+      const expected = BigFraction.fromString(testCase.quotient).toFixed(30);
       expect(result).toStrictEqual(expected);
     }
   });
@@ -57,16 +57,16 @@ describe("Arithmetic operations", () => {
   it("square root", () => {
     for (let i = 0; i < testCases.length; i++) {
       const testCase: TestCase = testCases[i];
-      const x = Fraction.fromString(testCase.x);
+      const x = BigFraction.fromString(testCase.x);
       if (!x.isNegative) {
         const result = x.sqrt().toFixed(30);
-        const expected = Fraction.fromString(testCase.sqrtX).toFixed(30);
+        const expected = BigFraction.fromString(testCase.sqrtX).toFixed(30);
         expect(result).toStrictEqual(expected);
       }
-      const y = Fraction.fromString(testCase.y);
+      const y = BigFraction.fromString(testCase.y);
       if (!y.isNegative) {
         const result = y.sqrt().toFixed(30);
-        const expected = Fraction.fromString(testCase.sqrtY).toFixed(30);
+        const expected = BigFraction.fromString(testCase.sqrtY).toFixed(30);
         expect(result).toStrictEqual(expected);
       }
     }
@@ -75,10 +75,10 @@ describe("Arithmetic operations", () => {
   it("square", () => {
     for (let i = 0; i < testCases.length; i++) {
       const testCase: TestCase = testCases[i];
-      const x = Fraction.fromString(testCase.x);
+      const x = BigFraction.fromString(testCase.x);
       // square
       const squared = x.square().toFixed(30);
-      const expectedSquare = Fraction.fromString(testCase.xSquare).toFixed(30);
+      const expectedSquare = BigFraction.fromString(testCase.xSquare).toFixed(30);
       expect(squared).toStrictEqual(expectedSquare);
     }
   });
@@ -86,14 +86,14 @@ describe("Arithmetic operations", () => {
   it("pow", () => {
     for (let i = 0; i < testCases.length; i++) {
       const testCase: TestCase = testCases[i];
-      const x = Fraction.fromString(testCase.x);
+      const x = BigFraction.fromString(testCase.x);
       // square
       const squared = x.pow(2).toFixed(30);
-      const expectedSquare = Fraction.fromString(testCase.xSquare).toFixed(30);
+      const expectedSquare = BigFraction.fromString(testCase.xSquare).toFixed(30);
       expect(squared).toStrictEqual(expectedSquare);
       // cube
       const cubed = x.pow(3).toFixed(30);
-      const expectedCube = Fraction.fromString(testCase.xCube).toFixed(30);
+      const expectedCube = BigFraction.fromString(testCase.xCube).toFixed(30);
       expect(cubed).toStrictEqual(expectedCube);
     }
   });
